breaking change of types and rethinking plotting
change `Geometry` and `GRoute` types but the functions using them have not
changed to use the types appropriately, this might break things if
run. There are bunch of TODO's so take of that if you have time.
Wanted to start doing the origin/destination point routing but i did
not get to start it, think i need task tracking or something. You
have to showcase this next week so break a leg mister!.