fixed getting plotting arguments
getting (default) keyword arguments to do plotting is now better than
the first approach. happy with how the whole thing evolved. took a
shot at starting the shortest path analysis part but i'm out of
flow and i'll fuck this code up if i try. as it stands now, the
whole thing is hanging on by a thread. the week is about to start
and I have not met any goals i set for last week so yaaaayyy...